+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/deploy-tooling
 
+## 0.15.25
+
+### Patch Changes
+
+-   Updated dependencies [3734fe8]
+    -   @sap-ux/btp-utils@0.16.0
+    -   @sap-ux/axios-extension@1.17.2
+    -   @sap-ux/system-access@0.5.17
+
 ## 0.15.24
 
 ### Patch Changes